awesome: implement function to swap clients between two screens
This is useful for clients that are on the smaller laptop screen to
monitor something but which need to be on the larger main screen to ease
reading. This function swaps the windows from the current active tag on
screen 0 with those of from the current active tag on screen 1.

The function is not well tested and there seem to be certain side
effects which are not fully understood.