awesome: add first iteration of thinkpad battery widget
Newer Lenovo ThinkPads (T400, X220, ...) no longer come with a battery
LED visible to the user if the lid is open. The only battery indicator
is now on the outside of the thinkpads and thus only visible if the lid
is closed.

This adds a primitive widget displaying an icon that is either green (if
the battery charge level is greater than 20 percent) or orange (if the
battery charge level is below 20 percent) icon and the estimated time in
'%H:%M' until the battery is depleted.

The icon is imitated after a photo which was made of the X220 battery
LED. The colors are not yet set in stone.