power_supply: rewrite functions for arbitrary number of batteries
Each battery icon is displayed gray and there is no estimated time on
startup. The first update currently is after the first timer timeout
which is 5 minutes.

A battery that is available but whose status is 'Unknown' is displayed
gray without estimate.

There is a quick 'n dirty approach to the energy_ vs. charge_ difference
which certainly needs to be reworked wrt. code reuse.

Both imagebox and textbox widgets are updated after 5 minutes using the
same timer.